Reorganize github fetchers
